DCW - Now supports CIDR syntax w/ Net::CIDR::Lite
add_range() now looks for a CIDR suffix like /24 on the start IP
address. If one is found, it uses a new dependency Net::CIDR::Lite to
convert that into a range itself. Any passed in end IP address is ignore
in this case.

Plus found bug in Set::IntervalTree with the exact IP address 128.0.0.0
which is 2^31. It will return a false positive for any non-empty tree.

Plus cleanup of test script.